[FIX] Add workarround for pthread_cancel that is not supported by Android
bionic, the android libc specificaly stated that pthread_cancel is not
to be supported :
https://android.googlesource.com/platform/bionic/+/ics-mr1-release/libc/docs/OVERVIEW.TXT

Therefore we had to replace pthread_cancel usage with something working
with bionic. We choose to use a signal based mecanism as SIGUSR1 is not
used by the rest of the library.

There is no system call error handling in the rest of the library and
function return values are ignored. I prefer to make my change minimal
so I dont fix it.